// lib/types.ts
export type TFStatement = {
  text: string;
  answer: boolean; // true = Đúng, false = Sai
  explain: string;
};

export type TFQuestion = {
  id: string;
  context: string; // lời dẫn tình huống (3 dòng trở lên, có ngữ cảnh sinh động)
  code?: string;
  statements: TFStatement[]; // luôn 4 ý
};